Rename: marketplace를 prompts 로 변경
