--- apps/client/src/app/marketplace/page.tsx
@@ -1,9 +1,0 @@
-import MarketPlaceTemplate from "@/components/marketplace/template/MarketPlaceTemplate"
-
-export default function page() {
-	return (
-		<section>
-			<MarketPlaceTemplate />
-		</section>
-	)
-}
