+++ apps/client/src/app/prompts/page.tsx
@@ -1,0 +1,9 @@
+import MarketPlaceTemplate from "@/components/prompts/template/PromptsTemplate"
+
+export default function page() {
+	return (
+		<section className="min-h-screen bg-[#111111]">
+			<MarketPlaceTemplate />
+		</section>
+	)
+}
