+++ apps/client/src/components/prompts/atom/prompts/PromptPrice.tsx
