+++ apps/client/src/components/prompts/atom/prompts/PromptThumbnail.tsx
