+++ apps/client/src/components/prompts/molecule/prompts/HoverPromptDetail.tsx
