+++ apps/client/src/components/prompts/template/PromptsTemplate.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import FilterSideBar from "../organism/FilterSideBar"
 import PromptList from "../organism/PromptList"
-import promptListData from "@/dummy/marketplace/promptListData"
-import { FilterData } from "@/types/marketplace/marketPlaceType"
+import promptListData from "@/dummy/prompts/promptListData"
+import { FilterData } from "@/types/prompts/marketPlaceType"
 
 function MarketPlaceTemplate() {
 	return (
